requirements.txt
django==1.6.1
gunicorn==18.0
django-debug-toolbar==0.11.0
django-taggit==0.11.2